convert-*.py: move per model weight estimation away from util back to main script
plus some refactoring
